ContentAutomator/Postify/requirements.txt
aiofiles==0.8.0
aiogram==3.0.0b3
aiohttp==3.8.4
aioschedule==0.5.2
aiosignal==1.3.1
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
facebook-sdk==3.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pydantic==1.10.7
Pygments==2.15.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
schedule==1.2.0
six==1.16.0
typing_extensions==4.5.0
tzlocal==5.0.1
urllib3==1.26.6
yarl==1.9.2
